+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v3.2.3-dev1] - 2023-11-06
 ### Fixed
 * Error in documentation example on drawing a time curve when mask is moved. Fixed.
+* Viewer: Updated MyPolygonSelector to match Matplotlib 3.8.
 
 ## [v3.2.3-dev0] - 2023-10-17
 ### Fixed